requirements.txt
pyppeteer==0.2.6
python-dotenv==0.20.0
aiohttp==3.8.1
aiosqlite==0.17.0
imaplib2==2.6.0
openai=1.14.1
